feat: course-detail
