+++ src/main/java/com/lxp/course/domain/model/CourseDetail.java
@@ -1,0 +1,6 @@
+package com.lxp.course.domain.model;
+
+
+public record CourseDetail(
+        String content,
+        String contentDetail) {}
