+++ src/main/java/com/lxp/course/persistence/row/CourseRow.java
@@ -1,0 +1,16 @@
+package com.lxp.course.persistence.row;
+
+import java.sql.Timestamp;
+
+public record CourseRow(
+        Long courseId,
+        String title,
+        Long instructorId,
+        Long categoryId,
+        int totalTime,
+        int totalLectureCount,
+        String content,
+        String contentDetail,
+        Timestamp createdAt,
+        Timestamp updatedAt) {
+}
