+++ src/main/java/com/lxp/course/persistence/row/SectionRow.java
@@ -1,0 +1,11 @@
+package com.lxp.course.persistence.row;
+
+import java.sql.Timestamp;
+
+public record SectionRow(
+        Long sectionId,
+        Long courseId,
+        String name,
+        int orderNum,
+        Timestamp createdAt,
+        Timestamp updatedAt) {}
