+++ src/main/java/com/lxp/course/persistence/row/TagRow.java
@@ -1,0 +1,10 @@
+package com.lxp.course.persistence.row;
+
+import java.sql.Timestamp;
+
+public record TagRow(
+        Long tagId,
+        String name,
+        Timestamp createdAt
+) {
+}
