--- src/main/java/com/lxp/domain/course/CourseDetail.java
@@ -1,8 +1,0 @@
-package com.lxp.domain.course;
-
-import java.time.Instant;
-
-
-public record CourseDetail(
-        String content,
-        String contentDetail) {}
