--- src/main/java/com/lxp/domain/course/CourseRepository.java
@@ -1,7 +1,0 @@
-package com.lxp.domain.course;
-
-import java.util.Optional;
-
-public interface CourseRepository {
-    Optional<Course> findAggregateById(Long courseId);
-}
