--- src/main/java/com/lxp/infrastructure/dao/course/SectionDao.java
@@ -1,41 +1,0 @@
-package com.lxp.infrastructure.dao.course;
-
-import com.lxp.infrastructure.mapper.SectionMapper;
-import com.lxp.infrastructure.row.course.SectionRow;
-import com.lxp.util.QueryType;
-
-import javax.sql.DataSource;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
-
-
-public class SectionDao {
-    private final DataSource dataSource;
-
-    public SectionDao(DataSource dataSource) {
-        this.dataSource = dataSource;
-    }
-
-
-    public List<SectionRow> findAllByCourseId(Long courseId) throws SQLException {
-        String query = QueryType.SECTION_FIND_ALL_BY_COURSE_ID.getQuery();
-        List<SectionRow> sectionRows = new ArrayList<>();
-
-        try (Connection connection = dataSource.getConnection();
-             PreparedStatement pstmt = connection.prepareStatement(query)) {
-
-            pstmt.setLong(1, courseId);
-
-            try (ResultSet rs = pstmt.executeQuery()) {
-                while (rs.next()) {
-                    sectionRows.add(SectionMapper.fromResultSet(rs));
-                }
-                return sectionRows;
-            }
-        }
-    }
-}
