+++ src/main/java/com/lxp/infrastructure/row/CategoryRow.java
@@ -1,0 +1,12 @@
+package com.lxp.infrastructure.row;
+
+import java.sql.Timestamp;
+
+public record CategoryRow(
+        Long categoryId,
+        String name,
+        Long parentId,
+        Timestamp createdAt,
+        Timestamp updatedAt
+) {
+}
