--- src/main/java/com/lxp/infrastructure/row/course/CourseRow.java
@@ -1,16 +1,0 @@
-package com.lxp.infrastructure.row.course;
-
-import java.sql.Timestamp;
-
-public record CourseRow(
-        Long courseId,
-        String title,
-        Long instructorId,
-        Long categoryId,
-        int totalTime,
-        int totalLectureCount,
-        String content,
-        String contentDetail,
-        Timestamp createdAt,
-        Timestamp updatedAt) {
-}
