--- src/main/java/com/lxp/infrastructure/row/course/SectionRow.java
@@ -1,11 +1,0 @@
-package com.lxp.infrastructure.row.course;
-
-import java.sql.Timestamp;
-
-public record SectionRow(
-        Long sectionId,
-        Long courseId,
-        String name,
-        int orderNum,
-        Timestamp createdAt,
-        Timestamp updatedAt) {}
