--- src/main/java/com/lxp/infrastructure/row/course/TagRow.java
@@ -1,10 +1,0 @@
-package com.lxp.infrastructure.row.course;
-
-import java.sql.Timestamp;
-
-public record TagRow(
-        Long tagId,
-        String name,
-        Timestamp createdAt
-) {
-}
