+++ src/main/java/com/lxp/util/QueryType.java
@@ -9,13 +9,14 @@ public enum QueryType {
 
     // course_queries.xml
     COURSE_FIND_BY_ID("course.findById"),
-
+    COURSE_FIND_BY_ID_JOIN_INSTRUCTOR("course.findByIdJoinInstructor"),
+    COURSE_FIND_BY_ID_JOIN_CATEGORY("course.findByIdJoinCategory"),
 
     //section_queries.xml
     SECTION_FIND_ALL_BY_COURSE_ID("section.findAllByCourseId"),
 
     //lecture_queries.xml
-    LECTURE_FIND_ALL_BY_SECTION_IDS("section.findAllBySectionIds"),
+    LECTURE_FIND_ALL_BY_SECTION_IDS("lecture.findAllBySectionIds"),
 
     //tag
     TAG_FIND_ALL_BY_COURSE_ID("tag.findAllByCourseId"),
