+++ src/main/java/com/lxp/util/QueryUtil.java
@@ -17,6 +17,7 @@ public class QueryUtil {
 //        loadQueries("queries/category_queries.xml");
         loadQueries("queries/course_queries.xml");
         loadQueries("queries/section_queries.xml");
+        loadQueries("queries/lecture_queries.xml");
         loadQueries("queries/tag_queries.xml");
 //        loadQueries("queries/enrollment_queries.xml");
         loadQueries("queries/tag_queries.xml");
